refactor: Resource file structure
Moved around a few things, to show clear distinction between what resources we manage an what are managed from other modules
