+++ src/components/GenerateReceivingModal/GenerateReceivingModal.test.js
@@ -17,10 +17,10 @@ import GenerateReceivingModal from './GenerateReceivingModal';
 
 import { translationsProperties } from '../../../test/helpers';
 import {
-  holdings as mockHoldings,
-  locations as mockLocations,
   pieceSet,
   serial,
+  holdings as mockHoldings,
+  locations as mockLocations,
 } from '../../../test/resources';
 
 const INTERVAL = 0;
@@ -83,7 +83,7 @@ jest.mock('react-query', () => {
 
         // Ensure we return the promise resolve from above, so that any _subsequent_ .then calls can flow
         return mockMutateAsync(...incomingParams);
-      }
+      },
     })),
   };
 });
@@ -202,7 +202,9 @@ describe('GenerateReceivingModal', () => {
   describe('filling out required fields', () => {
     beforeEach(async () => {
       await waitFor(async () => {
-        await TextField('Time between publication and receipt (days)*').fillIn('0');
+        await TextField('Time between publication and receipt (days)*').fillIn(
+          '0'
+        );
         await Select('Location*').choose(mockLocations[0].name);
       });
     });
