+++ test/resources/externalResources/holdings.js
