+++ test/resources/externalResources/locations.js
