+++ test/resources/handlers.js
@@ -1,3 +1,4 @@
+// Part of me thinks these dont really need to exist
 const handlers = {
   onClose: jest.fn(),
   onDelete: jest.fn(),
