+++ test/resources/index.js
@@ -1,9 +1,11 @@
-export { default as mockRefdata } from './refdata';
-export { default as handlers } from './handlers';
 export { default as serial } from './serial';
 export { default as pieceSet } from './pieceSet';
 export { default as ruleset } from './ruleset';
-export { default as locations } from './locations';
-export { default as holdings } from './holdings';
-export { default as orderLine } from './orderLine';
+
 export { default as locales } from './locales';
+export { default as mockRefdata } from './refdata';
+export { default as handlers } from './handlers';
+
+export { default as holdings } from './externalResources/holdings';
+export { default as locations } from './externalResources/locations';
+export { default as orderLine } from './externalResources/orderLine';
